Moved feed methods in collector to separate file
